Moved `common-rendering` lib functions to AR lib
