+++ apps-rendering/src/components/Layout/NewsletterSignUpLayout.tsx
@@ -3,7 +3,7 @@
 import type { SerializedStyles } from '@emotion/react';
 import { css } from '@emotion/react';
 import { background } from '@guardian/common-rendering/src/editorialPalette';
-import { maybeRender } from '@guardian/common-rendering/src/lib';
+import { maybeRender } from 'lib';
 import type { ArticleFormat } from '@guardian/libs';
 import {
 	brandAlt,
